Removed MovementRestriction interface

+++ src/engine/movements/KnightMovement.java
@@ -5,8 +5,8 @@ import engine.Coordinates;
 public class KnightMovement implements Movement {
 	@Override
 	public boolean canMove(Coordinates<Integer> from, Coordinates<Integer> to) {
-		int xDiff = (int) Math.abs(from.x() - to.x());
-		int yDiff = (int) Math.abs(from.y() - to.y());
+		int xDiff = Math.abs(from.x() - to.x());
+		int yDiff = Math.abs(from.y() - to.y());
 
 		return xDiff == 2 && yDiff == 1 || xDiff == 1 && yDiff == 2;
 	}
